Move `uncurry's to `Foundation.Tuple', add `curry'
